Live View: Update script from param with editor closed
Summary:
before it was only updating if you had the editor open.
now, it will update even if the editor is not open. it will also execute properly (there was a race condition before).
there is some duplicated code here, because we need to decide if we want to move the example scripts back out into the context, since it is not only used in the dropdown now.

Test Plan: tried it on webpack

Reviewers: zasgar, malthus, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4115

GitOrigin-RevId: 3da11deb6b15502ab4c86e9e4a329014b7797fb4