Fix prevResourceVersion bug in vizier metadata subscriber
Summary:
before, the first ever update that a vizier sends to cloud has a prevResourceVersion of "".
this worked fine before, when requests for missing updates were sent in one whole, big message. but, this message can get too large to send over NATS, so we need to start batching the missing metadata updates.
this meanas that the first ever update that a vizier sends to cloud should have the correct resource version, so that cloud can actually know when its received all of its missing updates.

Test Plan: update unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4123

GitOrigin-RevId: e8fe502dbdb3552fc3467a7166638e63f67aaa7d