Set prevResourceVersions in missing metadata update response
Summary:
when cloud makes a request for missing metadata updates, we call GetMetadataUpdatesForHostname. We werent actually setting the prevResourceVersions in these updates before, since we were just sending a single, large batch. now that we'll be having the missing update response send multiple batches, we'll need that information to piece everything together.

there was also a bug where we were sending extra updates.
take for example a podUpdate, with RV X. This is saved under /resourceVersionUpdate/X in etcd. however, this single update consists of multiple updates if the pod has containers. for two containers, there will be update X_0 for the first container, X_1 for the second, and X_2 for the actual pod update. when we request missing updates for X_1 to Y in this function, we actually make a call to etcd for versions X to Y. since resource update X actually consists of updates X_0, X_1, and X_2, we need to filter out X_0, since that wasn't in the requested range.

Test Plan: unit test, ran skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4127

GitOrigin-RevId: 45ca61d973d02b3549be1c97af4de70ce1f13ab1