Update vzmgr metadata syncing to listen to updates for newly connected viziers
Summary:
currently the metadata listener is not hooked up to listen to any vizier updates. this diff updates it so that we listen for any viziers that have just connected. when vzmgr receives a connected vizier, it sends a message over a NATS channel. we now listen to that channel and subscribe to the proper vizier update channels.
a followup diff will update the metadata listener so that it also starts listening for vizier updates for viziers that were already connected.

Test Plan: ran skaffold and verified it works... played around with unit tests for awhile, but found that it was pretty hard to actually test. since the vizier subscription is async, there doesn't seem to be a clean way of waiting and knowing when we've successfuly subscribed to the vizier's update NATS channel.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4133

GitOrigin-RevId: 7dfd93a875045fbce7082e94819b6ae406c16a34