PL-1692: Move load tester for sockshop from extensions/v1beta to apps/v1
Summary:
As of k8s v1.16 extensions/v1beta for Deployment is deprecated.
Deployment is apps/v1 for v1.12 which is the oldest version of K8s we support.
This should fix the issue that customers on higher versions were running into with deploying the demo.

Test Plan: try with kubernetes version 1.16 and above (1.17 to match the customer who filed the bug)

Reviewers: zasgar, michelle, #engineering, philkuz

Reviewed By: #engineering, philkuz

Subscribers: ishan

JIRA Issues: PL-1692

Differential Revision: https://phab.corp.pixielabs.ai/D4143

GitOrigin-RevId: 40ff646475f7afacdf257686c773c903685897ef