CLI deploy: save K8s cluster info to cluster secrets
Summary:
We want to send data about the user's cluster when Vizier registers with the cloud.
Some of this information, such as cluster name, will be hard to get from within the cloudconnector.
instead, when the user deploys through the CLI, we can load some of these variables into the cluster secrets:
- cluster name
- cluster k8s version
- vizier version (alternatively, we could also get this from within the cloudconnector pod, by parsing a vizier pod's image tag)

Test Plan: ran CLI and checked secrets

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4138

GitOrigin-RevId: 351da7cc1f33d854ebe3ba51b8df46b05beaa729