Uprobe deployment: Avoid duplicate ElfReaders
Summary:
Previously, the ElfReader was created twice for each binary:
 - As part of GetSymAddr()
 - As part of ResolveUprobeTmpls()

This diff refactors the logic so that the main loop is over the binaries and creates the ElfReader once.

The ElfReader is then passed as an argument to the UpdateSymAddrs and Probe deployment functions.

During the refactor, some logic was moved:
1) SymAddrs were highly HTTP2 specific, so pulled them out of elf_tools, which should be a generic library.
2) ResolveUprobeTmpls was pulled into SocketTraceConnector because the main binary iteration loop needs to be in SocketTraceConnector.

Test Plan: Existing tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4166

GitOrigin-RevId: 633edbbe19f54848d10c091e1814b57a5690d158