Fix metadata syncing bug
Summary:
when the metadata service starts up for the first time, it starts up a k8s listwatch for each resource. this gets some old resources which should have already been sent to cloud before. if not, these updates will be sent when cloud requests for missing metadata updates.
however, we were sending them anyway, and assigning them a prevResourceVersion based on the last resourceVersion that was sent out.. making the prevRV > RV, which confused the cloud metadata reader. we should just not send out these out-of-order updates.

Test Plan: ran in skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4181

GitOrigin-RevId: f53f41da200e86fae212615935302f1cc40de82a