Stirling: OpenSSL tracing
Summary:
This diff adds OpenSSL tracing to Stirling.

It has been structured to push data to Stirling in as similar a way as the kprobe data,
to minimize the amount of code divergence.

In BPF the conn_info_map entry will be updated to trace only SSL traffic once SSL is observed on the FD.

In user-space, a new connection tracker will be created for the SSL traffic.

Test Plan:
Manually tested on nginx.
Working on a containerized test in a separate diff.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1369

Differential Revision: https://phab.corp.pixielabs.ai/D4131

GitOrigin-RevId: 92d60ecae8c2e762095a82ba244178bd8e20cab5