ContainerRunner: Ability to run local tarballs
Summary:
This adds the ability for ContainerRunner to run local tarballs (instead of pulling images from a registry).
Useful when the local image is created by bazel.

In these cases, there is no need to push the image to a repo; we can just have bazel output the tarball,
and then have the test load it.

To be used by OpenSSL tests. Should be useful in other future cases as well.

Test Plan: This is a test utility, so no test.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4212

GitOrigin-RevId: 37ff503b254a1438579e43ecad8ba7a0d792c671