Add basic pod resource usage to pod_stats Live View
Summary:
TSIA. FYI that these new timeseries are also affected by PP-1757,
(results go to 0 for counter-based metrics as time window gets close to 1s).
This problem also exists in the current node_stats view (which the pxl added was
based on), so we can decide whether or not to delay merging this until PP-1757
is addressed.

Test Plan: ran it

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Subscribers: ishan

Differential Revision: https://phab.corp.pixielabs.ai/D4223

GitOrigin-RevId: 4e4b5665507fd6b2068c9e4436726dcb21132e97