stirling_perf_test script
Summary:
This test checks the CPU usage of Stirling. It's not meant to be a tight bound, but to catch egregious CPU usage increases.

Setting the limit to 10% to start with. We can adjust as necessary.

Depends on D4215.

Test Plan: Let's see how flaky it is on Jenkins with a 10% bound.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1749

Differential Revision: https://phab.corp.pixielabs.ai/D4218

GitOrigin-RevId: 04acca953f97047a1ed07f2b42583dd74e6514c1