PC-93 part 1: Set the stage for supporting the new Live View format in the UI
Summary:
We have the old Live View format and the new one, which is added in D4151.
Here is the current plan for the transition between them, which aims to do it in steps so that
it can get out the door faster and not be blocked on other initiatives:

1. Add support in the UI for the new Live View format, while temporarily maintaining support for the Old Live View format.
2. Add new Live View format to the script bundle, with a vis.json file that can be displayed by the UI.
3. Move current live views in the script bundle over to the new format.
4. Deprecate support for the old Live View format in the UI.
5. Have the UI move from storing the vis spec as json to storing it as protobuf for increased correctness.
6. Replace the script bundle with the script manager. (might be combined with 5).

This diff doesn't actually do much, it just starts to carve up the territory of the old way and add some handlers
for the new way. However, for the sake of smaller diffs I put it out, it just maintains parity with the current
feature set.

Test Plan: ran the live view UI with a Vizier

Reviewers: malthus, philkuz, zasgar, #engineering

Reviewed By: malthus, #engineering

JIRA Issues: PC-93

Differential Revision: https://phab.corp.pixielabs.ai/D4237

GitOrigin-RevId: 2e769cb935aea8200689125922cc58c3c9d59d9d