Add elastic to cloud API service
Summary:
we will need an elastic client in the API service, since that is where we will be handling the autocomplete requests.
this diff adds the client to the API service, but does not use it yet.

moved the connection code for elastic into utils, so that we dont have to copy & paste every time.

Test Plan: ran skaffold and made sure both API and log collector services start up without errors

Reviewers: zasgar, jamesbartlett, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4235

GitOrigin-RevId: 6fc8d933ea1736a4cb05bca1b08ff7bfe2585fc2