Add ebnf parser for autocomplete input
Summary:
we will use an ebnf parser to parse the user's autocomplete input into an action and args.
this will save us the trouble of manually parsing their input using strings.split... which was very ugly.
the next step is to iterate over the args, do some bookkeeping to see what entity kinds we should be searching for, then doing the actual search for suggestions

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4239

GitOrigin-RevId: 1c8d54ef989848b76ac8017c6317bc37f03c42b5