PC-102 Update login UI to take org_name as param
Summary: when logging in with a pixie service account, we need a way to specify which org they should be logging into. until we have better rbac rules, we will specify the org name upon login in a query param. this param should be passed in the login POST request to Pixie cloud

Test Plan: ran in skaffold, made sure that it does pass org_name to the api service when logging in

Reviewers: malthus, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4256

GitOrigin-RevId: 0e0d514c784e1f7fc0f9f459e5b450dbe2aa71e1