PC-107 Add orgname to user resolver
Summary:
When logged into the pixie support account and looking at another org's cluster, we want to show a banner notifying the user that they are viewing an external org.
to do that, we should be able to request the orgname from the UI.
the next step is to actually make the gql call in the UI, check the email domain, and the orgname, and show the banner if they don't match.

Test Plan: unit test, ran a graphql request for user and made sure orgname shows up

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4262

GitOrigin-RevId: f43736ea8236424830a2942ee752a131b6b5fdb2