Initial version of data-table.
Summary:
Rewrite of data table using react-virtualize grid.
Features include:
- Compact mode
- sticky header
- auto column width to fit the column content
- cell content alignment

The limitation for this version is that the table will not fill the width of the available size, but what it does allow is horizontal scroll and each column can grow beyond the available space until it is truncated.

Test Plan: storybook

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-99

Differential Revision: https://phab.corp.pixielabs.ai/D4272

GitOrigin-RevId: 20077cc379dc894f5b7a0bc2642ce90e091faeb5