Improved data stream parsing on mis-aligned cases
Summary:
If the head of events is not on a message boundary, all subsequent events until the next gap are flushed.
This is not ideal, since a single missing event may cause such cases to trigger, and we would lose a lot of data.

This diff makes the parsing more robust to such cases.

Test Plan: Modified existing tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1772

Differential Revision: https://phab.corp.pixielabs.ai/D4258

GitOrigin-RevId: bb116d231335758cb5bddd251028015b5b74a2d2