Autocomplete: Get formatted output string from parsed command
Summary:
This diff adds the function for constructing a formatted output string from a parsed command.
the formatted output string depends on the current cursor position and the action the user has taken. for example, if the user performs a select, then we should move their cursor to the next invalid tab stop. if they just did an edit, we should keep their cursor in the same position.
we should add the cursor marker ("$0") in the correct index in the string that is passed to ParseIntoCommand, so that it is in the correct TabStop. However, we should remember to not include it when requesting suggestions.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4273

GitOrigin-RevId: 2fe66e6a9060165f5aacfbcc6719e456d144c38e