ClientServerSystem: Consolidate model across Read/Write syscalls
Summary:
This cleanup is in preparation for the change in req-response matching.

In the upcoming diff, unmatched requests/responses are not pushed to the table,
so a lot of tests need updating.

This is prepatory work for that change.

Test Plan: This is just a change to test code, so rely on existing tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1763

Differential Revision: https://phab.corp.pixielabs.ai/D4268

GitOrigin-RevId: ee70e499d58375d8d0ed232a7108b7c6f675c7fb