Add arc unit test for checking pb breaking changes
Summary:
Adding unit test for prototool to check for any API breaking changes.
the test is currently not enabled, because i need to update chef to include prototool.

the ugliest part is that we have to copy any vendor module protos to a tmp directory with the correct structure/naming. but after some thought, this didnt seem so bad since we only use gogoproto.

at some point we can enable the proto linter if we want, but this only uses the `prototool breaking changes` tool.

Test Plan: ran it locally with some api breaking changes to make sure it complains.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4078

GitOrigin-RevId: 8ebb5462da8f3876470dd4b12486dab071da43e6