PC-92: Support new functional vizier ExecuteScript API
Summary:
the changes to the generated js/ts files will go away, because I will wait to land this until
@jamesbartlett lands his changes that is responsible for updating those.

The new version of Live View will execute scripts by function, and it added some fields to the grpc
API to support that. This diff adds that field as an optional argument for the vizier ts GRPC client
executeScript function, so that users of the GRPC client can call the new version of the api.

Test Plan: will test once the new api is landed

Reviewers: malthus, philkuz, zasgar, #engineering

Reviewed By: malthus

Subscribers: jamesbartlett

JIRA Issues: PC-92

Differential Revision: https://phab.corp.pixielabs.ai/D4231

GitOrigin-RevId: 7dddb6711d06030696192b8514f275faf680216f