Autocomplete: make multisearch request to elastic
Summary:
when making an autocomplete request, we generally make N requests to elastic, where N is the number of tokens. we can batch our requests to elastic in a multisearch request, so we only need to make a single call to elastic.
before, the average response time was 87.69ms. with multisearch, the avg response is time 66.37ms.

Test Plan: unit test, ran skaffold and made a gql request from UI

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4325

GitOrigin-RevId: 4c9861c9351971699019d07b35e452485208d45b