Fix Metadata list bug with > 500 items
Summary:
the k8s lister returns an internalversion.List object instead of a <Resource>List object when it exceeds 500 items.
this is probably because it needs to do some pagination behind the scenes. it ends up the internalversion.List object just contains a list of runtimeObjects in its Items field, which can be casted directly into the resource we are looking for.
Unfortunately, go doesn't support templates like C++ does, so we need to repeat the same code a few types for each different type...

Test Plan: ran in a cluster with 500 pods

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4328

GitOrigin-RevId: 82a728ed90658f5a6beb37125fe9645931a0db3c