PC-128 Add graph widget to live view
Summary:
We want to add a graph widget to the live view, which is basically what we used to implement the graph for explaining query execution in console view.
we first just want to support the case where the data is a single column, with a single rowbatch, containing the dot file to render.

next, we want to support the case where a specifies two columns to be an adjacency list for their graph, generate the dot file from that, and render that.

please let me know if im missing a step in adding a new widget!

Test Plan:
ran UI in webpack, updated vis spec to be
```
    {
      "name": "__query_plan__",
      "position": {
        "x": 0,
        "y": 9,
        "w": 9,
        "h": 4
      },
      "func": { // This is a random function because when #pl:set explain=true is on, it will always attach the query_results table to the output.
        "name": "resource_stats",
        "args": [
          {
            "name": "start_time",
            "value": "-30s"
          },
          {
            "name": "pod_name",
            "value": "/"
          }
        ]
      },
      "displaySpec": {
        "@type": "pixielabs.ai/pl.vispb.Graph",
        "dotColumn": "query_plan"
      }
    },
```
and check that the graph view is rendered in the live view
screenshot: {F66865}

Reviewers: nserrino, philkuz, jamesbartlett, malthus, zasgar, #engineering

Reviewed By: nserrino, malthus, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4344

GitOrigin-RevId: 4fd06a85a118725a35830f64682abe355711537a