Use (nodename, IP) pair to identify agents in metadata
Summary:
it is possible for nodename to not be unique, so we can't use it as a 1-1 mapping to agent.
instead, we should use the (nodename, IP) pair.

Test Plan:
ran skaffold... was able to run queries normally with servicenames/podnames resolved...
but probably additional testing on EKS needs to be done

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4357

GitOrigin-RevId: dd17f0363b18e9bcf957d36bc5c2efec7040848c