BPF: Make verifier happy on kernel 4.15.18
Summary:
We had BPF verifier regressions because kernel 4.14 was no longer being checked by Jenkins.

This is at least one piece needed to get back to functionality on 4.14.

Prior to this change, 4.15.18 kernels would not deploy. This fixes it.

Next up is to check 4.14 kernels.

Test Plan: Ran it. Need to restore Jenkins workers that use 4.14 kernel, but that can happen separately.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-1786

Differential Revision: https://phab.corp.pixielabs.ai/D4365

GitOrigin-RevId: fd76b09196fccd7c939ebd25f9fc597a82bf7c62