Update renew period for Cloud connector
Summary: Renew period is used to describe how long to wait before renewing a leader lease. It also is converted into a retry period by dividing renew period by 4. I think the the retry period on each command is too short, meaning a normal latency request to Update the leader election has a high chance of failing, leading the campaign to look like it failed to update.

Test Plan: Tried this out on my gke cluser and have had 0 failures in an hour. Will continue to monitor and update if this changes

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4370

GitOrigin-RevId: 0175c875e0c03018bc4e877cacc1a2bf9cd674f8