Fix create_release script CLI bazel target
Summary: the create_release script uses the bazel target to determine how many files have changed since the last release. it appends this number to the end of the release candidate tag (0.1.29-pre-master.0). Noticed that the CLI rcs consistently had a .0 at the end, which wasnt right. at some point we changed the target to :px. updating this fixed the value

Test Plan: ran the script, made sure it had the correct #

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4377

GitOrigin-RevId: 9766adb8c8bead305202c9c6593e76907b71ebd2