Uprobe compatibility with 4.14 kernel
Summary:
The 4.14 kernel's BPF verifier is finicky, because it doesn't like bpf_probe_reads with a size of zero.

This diff changes all bpf_probe_reads that have a variable size to read an extra byte, to keep the verifier happy.

To accomodate the extra byte when the size is already at its max value, an extra byte is also added to the destination structs.

Test Plan: Existing tests + manual run on 4.14.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1791

Differential Revision: https://phab.corp.pixielabs.ai/D4381

GitOrigin-RevId: 54dd7c79913a0d4c2c056b57bc7c87779af0b577