Have localstorage updates to placement and vis only be triggered by updates to their state
Summary:
There were some weird sequences of actions that could lead to a weird state in the old live view.
Basically if the local storage and the state for placement or vegaSPec were updated around the same time,
there was a race condition. This diff makes it so that the local storage is only updated once the state
is updated, so that there is not a race condition. This does have the implication that if the user
types in a bad spec, the erroneous version won't be persisted on refresh.
We need to be sure that is how we want to handle bad specs before we land, but we have seen other errors
from a bad spec causing state issues in the UI so IMO it's a reasonable tradeoff to consider.

Test Plan: ran the UI with a bunch of cases with editing scripts, using the command module, etc.

Reviewers: malthus, michelle, zasgar, #engineering

Reviewed By: malthus, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4421

GitOrigin-RevId: cad259e3b5bd642eadb04175b6749e0a14cc90f4