HTTP2 Tracing: Support alternate golang http2 implementation
Summary:
Turns out there are two HTTP2 implementations in golang.

One is in the dedicated HTTP2 library.
The other is in the HTTP library.

There appears to be some copy-paste between the two implementations, but there are still significant differences.

For this reason we need alternate probes to attach to the HTTP library.

Tried to re-use as much of our existing probes.

Test Plan: None yet. Will add test before enabling the probes. This is just to check that old behavior is not broken.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1751

Differential Revision: https://phab.corp.pixielabs.ai/D4383

GitOrigin-RevId: fb7482c0477e0fe6cd43ae37ada8515bab3dabfe