Jenkins: Use 4.14 kernel
Summary:
This diff adds a new WORKER_NODE into the Jenkinsfile, that will enable use to switch Jenkins worker images.
This is used to switch us back to a new (but older) 4.14 kernel.

In the future, when we create new images for nodes, we can use the WORKER_NODE to make more seamless transitions to the new nodes.

Test Plan: Jenkins

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-1795

Differential Revision: https://phab.corp.pixielabs.ai/D4417

GitOrigin-RevId: 5c087ce1cd5ca68e82e5322669b451de337f07f3