PX-12 add search box
Summary:
This adds search in the same format as less. / activates the search which does a real time search
until you hit enter. At which point "n" can be used to find the next match. Backspace/delete can
be used to update the search string.

Escape exits search.

This is the last diff of PX-12 and closes the issue.

Depends on D4431.

Test Plan: Tested extensivel manually.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PX-12

Differential Revision: https://phab.corp.pixielabs.ai/D4432

GitOrigin-RevId: 85219035ef74d45532dd33632948bf20312ca9fb