Golang HTTP2 uprobes: Probe writeFrame for server responses
Summary:
Another critical piece of getting tracing of HTTP2 on the net/http golang library.

This one traces header writes from the server.

This probe is more complex, because it is actually multiple probes that are connected to each through context.

Depends on D4451.

Test Plan: Manually tested for now. But code is not enabled yet anyways. Test to come in a separate diff.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1751

Differential Revision: https://phab.corp.pixielabs.ai/D4452

GitOrigin-RevId: 0dda061435af1a8a149f7aba14529617c115c6b8