CLI: update all uses of kubectl delete to use k8s api delete
Summary:
we are removing our dependencies to kubectl. we use it a lot to delete resources.

there is a fair bit of copying and pasting in here, but there are a few minor differences, such as whether namespace is specified and the k8s API we should be using, which depends on the resource type. so i'm not sure if its worth it to use genny for this

Test Plan: ran deploy/delete a bunch of times

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4462

GitOrigin-RevId: caa7ce7f06a86f47a72d0a1a1ecf0b8368713de3