Add RPC call for updating vizier cluster
Summary:
we need to make an rpc call from the CLI to cloud to notify cloud that a Vizier cluster should be updated.
this request should make a call to vzmgr, which will shoot off a NATS message to the correct vizier that it should update.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4464

GitOrigin-RevId: b191717de0b9c106f59c3dc4aafae4bd8328bca7