[PP-1707] Remove artifacts on boundaries of Vega script
Summary:
Dependent on D4324. Removes the artifacts on the boundaries of Vega visualizations. Add two transformations to the data used by Vega to create timeseries charts, seems to clean up a lot.
Will have to wait for the tooltip UI to land before i can land this.

Test Plan: Tested on a local version of the UI.

Reviewers: malthus, jamesbartlett, nserrino, #engineering

Reviewed By: malthus, #engineering

JIRA Issues: PP-1707

Differential Revision: https://phab.corp.pixielabs.ai/D4356

GitOrigin-RevId: 70281eb6a90b3bbf1ebbb3550b6ffc35ad615c47