PL-1809: Add a bloomfilter protobuf message
Summary:
We are going to be using bloom filters to prune queries from running on agents that don't
contain data for a particular query (for example, don't run on an agent not monitoring pod foo when
we only filter down to data from pod foo in the query). This diff adds the bloomfilterpb reperesentation
which will be the wire format as the bloom filter is sent from the agent to the metadata service.
The metadata service will then provide the set of bloom filters (one per agent) to the query broker
which will pass it down to the compiler, where they can then be used for pruning.
Depends on D4658.

Test Plan: n/a here

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1809

Differential Revision: https://phab.corp.pixielabs.ai/D4477

GitOrigin-RevId: 92319d9267121efba2695a87a1333a0d544fc41a