Add dual locking to agent queues
Summary:
before we had a single lock on the entire array of agent queues. everytime we received an update or wanted to pull an update from an agentqueue, we would grab the whole lock.
now, we should grab that lock just to get a handle on the specific agent queue, then grab the lock specifically for that agent queue.

Test Plan: ran skaffold, agents seem to be heartbeating and responding to queries as normal. however, all i am running is sockshop.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4484

GitOrigin-RevId: 7a035606f02866069c9b74eefaea7041075f82b9