Golang HTTP2/GRPC Probing: Enable
Summary:
This diff enables support for tracing the Golang net/http implementation of HTTP2.

A lot of the work was already landed in earlier diffs. This mostly enables the tracing, and adds a test.

Test Plan: A containerized test using our a client-server built with our pixie framework is added.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1751

Differential Revision: https://phab.corp.pixielabs.ai/D4471

GitOrigin-RevId: 68865c07c77588808d8de4450234d1ea29b6d792