Reduce MySQL mis-classifications
Summary:
The StmtSendLongData command often parses as MySQL, because the command is quite general. It can have virtually any length, and it has no response to help further filter out bad classifications.

This diff adds a state to the MySQL connections that tracks the confidence. The variable 'active' starts off in false state and is only flipped on when we've seen a correctly parsed complex command.

This gives us more confidence that the connection is indeed MySQL, and we start pushing out records from that time onward.

Test Plan: Added a test case collected from GKE.

Reviewers: yzhao, chengruizhe, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-1750

Differential Revision: https://phab.corp.pixielabs.ai/D4498

GitOrigin-RevId: df9ba1e20db5bbc31ba3125e2ee06311a689d7ec