PC-177: Add UI support for binned bar charts (histograms)
Summary: TSIA. when making a bar chart with numeric data in the x axis (aka a histogram) then the vega options need to be set slightly differently. In the future we should infer the column type from the schema we receive in the UI but for now we can support numeric histograms only.

Test Plan: ran it with a new live view edit using it

Reviewers: malthus, philkuz, zasgar, #engineering

Reviewed By: malthus, #engineering

JIRA Issues: PC-177

Differential Revision: https://phab.corp.pixielabs.ai/D4524

GitOrigin-RevId: 761635ffa7ea938218af56efb8dc59242dc9d4d6