PC-70 pt 2: Adding esutils.Index to handle creation/updating of elasticsearch indices.
Summary:
This diff adds the Index object which handles creation of indices, as well as updates.
It attempts to get the elastic index to conform to the spec its passed.
Note that currently reindexing is not implemented, so only changes that don't need a reindex are supported.

Test Plan: Added tests

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-70

Differential Revision: https://phab.corp.pixielabs.ai/D4535

GitOrigin-RevId: 48b9a64c3aea25449b70f8b883264970ef1dde52