Add container test for pgsql
Summary:
It's not straightforward to get the target pid, because the psql commands runs to finish; which
leaves no time for docker inspect to find the pid.

Expectations are verified against the request text in the record batches.

Also, stirling capture traffic from both client and server. Because the client's remote address is
considered out of cluster.

One additional change is move PrintRecordBatch() to output.h, that's used for debugging, I think
it's worth to be reused elsewhere.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4505

GitOrigin-RevId: f47b3f2c7772479473a966550351e9b060ee9963