On UpdateOrInstallCluster request, send update message to Vizier through NATS
Summary:
when we initiate an UpdateOrInstallCluster rpc call from the CLI, the api service should make a request to vzmgr, which sends a nats message to the vizier in question.
it sends a response back once the vizier is able to confirm that it has received the update request.

this diff only includes the cloud changes. changes for cloudconnector to listen to the update request will come in a later diff

Test Plan: unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4478

GitOrigin-RevId: e9f3f7668dee9d6aa32f67bc5f2e5758faabefb0