Add maintenance mode for viziers in vzmgr
Summary:
when a vizier is updating, it should be put into maintenance mode. a vizier is in maintenance mode, it will not be set to disconnected even if it misses a heartbeat.
the vizier is put in maintenance mode when an update request successfully goes through.
we will need to listen to the status stream from the updater job in order to determine when to take a vizier off maintenance mode.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4574

GitOrigin-RevId: 5eb8468430accbb7156062322a8c3ed5c98c46de