Update vzmgr to set status to "CONNECTED" on registration
Summary:
upon registration, vizier's status should be "CONNECTED", since we don't actually know if it is healthy or not.
to maintain backwards compatibility, we also need to send the status in the heartbeat. if no status is specified, we assume it is healthy.

Test Plan: unit test, started up skaffold and looks like everything is still working fine with staging

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4578

GitOrigin-RevId: 7e26921f7600fbf68c8367f632cd212924622e34