Cloud connector should check for updater job upon startup
Summary:
when cloudconnector starts up, we should check if there is an update job running. if there is, then we should wait for it to complete before we update the state to healthy.
if it has failed, then we want to mark the vizier cluster as UPDATE_FAILED status

Test Plan: n/a, its hard to unit test k8s api things

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4587

GitOrigin-RevId: b86ba1369ac60615fbec3664b665a05a4681db05