PX-32 - Initial implementation of 'px update'
Summary:
Update works with other change in branch. However, it does not always wait for the health check since there is a race between
running the update and waiting on the health check. The sleep right now works for most cases, but not all. Next step is to remove that flaky code.

Test Plan: Tested `px update vizier`

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PX-25, PX-32

Differential Revision: https://phab.corp.pixielabs.ai/D4597

GitOrigin-RevId: 624b3f7dc5d74a2aa88d70f5e728fd2410d9ff4e