Linux headers package update
Summary:
We suspect a mis-matched header may be causing Stirling issues when it has to use the prepackaged headers.

To address this, this includes a set of Linux headers build from version 5.3.18.

The headers aren't used yet. This just updates the generation code, and makes the tarball accessible.

Test Plan: Ran the script to build the headers. They look fine.

Reviewers: yzhao, #engineering, zasgar

Reviewed By: #engineering, zasgar

JIRA Issues: PP-1789

Differential Revision: https://phab.corp.pixielabs.ai/D4583

GitOrigin-RevId: efabc0bf50e37344f79e53da648d90c0db94cd02